test: complete Group B e2e → feature migration with strengthened assertions (#225)
Ports remaining Group B tests into the feature tier, re-routes cross-file
tests to their natural homes (definition/references/rename/symbols), and
deletes the 5 now-superseded e2e source files.

Deleted:
- e2e_selection_range.rs    → feature_folding.rs
- e2e_workspace_diagnostics.rs → feature_diagnostics.rs
- e2e_cross_file.rs         → feature_definition, _references, _rename, _symbols
- e2e_symbol_resolve.rs     → feature_symbols.rs
- e2e_file_rename.rs        → feature_rename.rs

Assertion improvements:
- selection_range: full monotonic parent-chain expansion check
- pull_diagnostics: pinned kind='full' and items=[] on first pull
- workspace_diagnostic: exact item count instead of non-empty
- goto_definition cross-file: line number pinned
- references cross-file: count >= 2 added
- symbol_resolve passthrough: end.line assertion added
- rename/file-rename: snapshot-pinned via expect!
