+++ tests/feature_rename.rs
@@ -2,7 +2,7 @@
 
 mod common;
 
-use common::TestServer;
+use common::{TestServer, canonicalize_workspace_edit};
 use expect_test::expect;
 
 #[tokio::test]
@@ -168,4 +168,110 @@ async fn rename_on_nonexistent_symbol_does_not_error() {
     s.open("rn.php", "<?php\n// nothing to rename\n").await;
     let resp = s.rename("rn.php", 1, 5, "NewName").await;
     assert!(resp["error"].is_null(), "rename errored: {resp:?}");
+}
+
+// --- psr4-mini fixture: cross-file rename + PSR4-aware file rename ---
+
+/// Set up psr4-mini with all three files open in the document store.
+/// Both the in-file rename and willRenameFiles handlers require open documents.
+async fn psr4_bring_up() -> TestServer {
+    let mut server = TestServer::with_fixture("psr4-mini").await;
+    server.wait_for_index_ready().await;
+    let (user, _, _) = server.locate("src/Model/User.php", "<?php", 0);
+    server.open("src/Model/User.php", &user).await;
+    let (reg, _, _) = server.locate("src/Service/Registry.php", "<?php", 0);
+    server.open("src/Service/Registry.php", &reg).await;
+    let (greet, _, _) = server.locate("src/Service/Greeter.php", "<?php", 0);
+    server.open("src/Service/Greeter.php", &greet).await;
+    server
+}
+
+/// Renaming `class User` to `Account` must rewrite every `use App\Model\User`
+/// import in dependent files. Snapshot-pinned so byte-offset regressions are
+/// caught immediately.
+#[tokio::test]
+async fn rename_class_edits_all_dependents() {
+    let mut server = psr4_bring_up().await;
+    let (_, line, ch) = server.locate("src/Model/User.php", "class User", 0);
+
+    let resp = server
+        .rename("src/Model/User.php", line, ch + 6, "Account")
+        .await;
+
+    assert!(resp["error"].is_null(), "rename error: {resp:?}");
+    let root = server.uri("");
+    let snap = canonicalize_workspace_edit(&resp["result"], &root);
+    expect![[r#"
+        // src/Model/User.php
+        4:6-4:10 → "Account"
+
+        // src/Service/Greeter.php
+        4:14-4:18 → "Account"
+
+        // src/Service/Registry.php
+        4:14-4:18 → "Account""#]]
+    .assert_eq(&snap);
+}
+
+/// Moving `src/Model/User.php` to `src/Entity/User.php` changes the FQN from
+/// `App\Model\User` to `App\Entity\User`; every `use App\Model\User` must be
+/// rewritten.
+#[tokio::test]
+async fn will_rename_file_rewrites_use_imports_in_dependents() {
+    let mut server = psr4_bring_up().await;
+    let old_uri = server.uri("src/Model/User.php");
+    let new_uri = server.uri("src/Entity/User.php");
+
+    let resp = server.will_rename_files(vec![(old_uri, new_uri)]).await;
+
+    assert!(resp["error"].is_null(), "willRenameFiles error: {resp:?}");
+    let root = server.uri("");
+    let snap = canonicalize_workspace_edit(&resp["result"], &root);
+    expect![[r#"
+        // src/Service/Greeter.php
+        4:4-4:18 → "App\\Entity\\User"
+
+        // src/Service/Registry.php
+        4:4-4:18 → "App\\Entity\\User""#]]
+    .assert_eq(&snap);
+}
+
+/// Renaming a file to the same PSR4-derived FQN must be a no-op.
+#[tokio::test]
+async fn will_rename_file_same_psr4_fqn_produces_no_edits() {
+    let mut server = psr4_bring_up().await;
+    let old_uri = server.uri("src/Model/User.php");
+    let new_uri = old_uri.clone();
+
+    let resp = server.will_rename_files(vec![(old_uri, new_uri)]).await;
+    assert!(resp["error"].is_null(), "willRenameFiles error: {resp:?}");
+    let changes = resp["result"]["changes"]
+        .as_object()
+        .cloned()
+        .unwrap_or_default();
+    assert!(
+        changes.is_empty(),
+        "rename-to-self must not produce edits, got: {changes:?}"
+    );
+}
+
+/// Deleting the file that defines `App\Model\User` must strip the `use` line
+/// from every dependent.
+#[tokio::test]
+async fn will_delete_file_strips_use_imports_from_dependents() {
+    let mut server = psr4_bring_up().await;
+    let uri = server.uri("src/Model/User.php");
+
+    let resp = server.will_delete_files(vec![uri]).await;
+
+    assert!(resp["error"].is_null(), "willDeleteFiles error: {resp:?}");
+    let root = server.uri("");
+    let snap = canonicalize_workspace_edit(&resp["result"], &root);
+    expect![[r#"
+        // src/Service/Greeter.php
+        4:0-5:0 → ""
+
+        // src/Service/Registry.php
+        4:0-5:0 → """#]]
+    .assert_eq(&snap);
 }
